+++ tests/client/test_workspaces.py
@@ -122,3 +122,39 @@ def test_workspace_delete_user(owner: "ServerUser") -> None:
 
     with pytest.raises(ValueError, match="Either the user with id="):
         workspace.delete_user(owner.id)
+
+
+def test_set_new_workspace(owner: "ServerUser"):
+    import argilla as rg
+
+    rg.init(api_key=owner.api_key)
+    ws = rg.Workspace.create("new-workspace")
+
+    rg.set_workspace(ws.name)
+    assert rg.get_workspace() == ws.name
+
+
+def test_init_with_workspace(owner: "ServerUser"):
+    workspace = WorkspaceFactory.create(name="test_workspace")
+
+    import argilla as rg
+
+    rg.init(api_key=owner.api_key, workspace=workspace.name)
+
+    assert rg.get_workspace() == workspace.name
+
+
+def test_set_workspace_with_missing_workspace(owner: "ServerUser"):
+    import argilla as rg
+
+    rg.init(api_key=owner.api_key)
+
+    with pytest.raises(ValueError):
+        rg.set_workspace("missing-workspace")
+
+
+def test_init_with_missing_workspace(owner: "ServerUser"):
+    import argilla as rg
+
+    with pytest.raises(ValueError):
+        rg.init(api_key=owner.api_key, workspace="missing-workspace")